section2/requirements.txt
flask
numpy
sklearn
Pillow
google-cloud-storage
tensorflow